+++ changelog.md
@@ -21,6 +21,8 @@ Codion Change Log
 - DefaultFrameBuilder bug fixed, now always makes sure the frame fits the screen.
 - DefaultComponentDialogBuilder, DefaultListSelectionDialogBuilder and DefaultComboBoxSelectionDialogBuilder bug fixed, now always make sure the dialog fits the screen.
 - DefaultFilterTableCellRenderer bug fixed, no longer sets tooltips for null column values.
+### is.codion.plugin.flatlaf
+- UIScaler bug fixed, now supports the codion intellij themes.
 ### is.codion.plugin.flatlaf.themes
 - Module is.codion.plugin.flatlaf.lookandfeels renamed is.codion.plugin.flatlaf.themes.
 ### is.codion.swing.framework.model